🐛 Use unsigned bitwise operands
Make shift widths and masks explicitly unsigned and handle the full signed exponent range without overflow.

+++ bindings/mlir/qiskit/QiskitImport.cpp
@@ -1912,7 +1912,7 @@ void translateCircuit(mlir::qc::QCProgramBuilder& builder,
       const auto arity = denseUnitaryArity(instruction);
       auto targets = std::span{operands}.subspan(arity.controls);
       std::ranges::reverse(targets);
-      const auto dimension = int64_t{1} << arity.targets;
+      const auto dimension = static_cast<int64_t>(uint64_t{1} << arity.targets);
       const auto type = mlir::RankedTensorType::get(
           {dimension, dimension}, mlir::ComplexType::get(builder.getF64Type()));
       const auto values = circuit.unitary(index);
